Fix mishandling of `--`.
With the current code, this invocation:
```
cargo llvm-lines -- --blah
```
results in a second invocation:
```
cargo llvm-lines -- --blah --filter-cargo
```
In this second invocation the `--filter-cargo` is ignored because it
comes after the `--`, so it executes normally rather than doing the
intended filter-and-terminate thing. This results in a third invocation:
```
cargo llvm-lines -- --blah --filter-cargo --filter-cargo
```
And so on, giving an infinite loop.

Inserting the `--filter-cargo` earlier, before the `--`, fixes the
problem.

The commit also adds a bunch of comments. I found this filtering stuff
quite puzzling and it took some time to work out what was going on.